ftmanager: Compiler delint

git-svn-id: https://svn.aros.org/svn/aros/trunk@49692 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/system/ftmanager/fontbitmap_class.c
@@ -330,7 +330,7 @@ IPTR fbNew(Class *cl, Object *o, struct opSet *msg)
 
 				if (!init)
 				{
-					int k;
+					ULONG k;
 					ULONG *p = colors;
 					for (k = 256; --k >= 0; p += 3)
 					{
